+++ app/src/main/java/net/maxproit/salesforce/masum/model/prospectmodel/CoApplicant.java
@@ -92,34 +92,34 @@ public class CoApplicant {
     private String netSalaryType;
     @SerializedName("netSalary")
     @Expose
-    private Integer netSalary;
+    private double netSalary;
     @SerializedName("businessIncome")
     @Expose
-    private Integer businessIncome;
+    private double businessIncome;
     @SerializedName("apartmentIncome")
     @Expose
-    private Integer apartmentIncome;
+    private double apartmentIncome;
     @SerializedName("semipakaIncome")
     @Expose
-    private Integer semipakaIncome;
+    private double semipakaIncome;
     @SerializedName("commercialSpaceIncome")
     @Expose
-    private Integer commercialSpaceIncome;
+    private double commercialSpaceIncome;
     @SerializedName("factoryIncome")
     @Expose
-    private Integer factoryIncome;
+    private double factoryIncome;
     @SerializedName("agriculturalIncome")
     @Expose
-    private Integer agriculturalIncome;
+    private double agriculturalIncome;
     @SerializedName("tutionIncome")
     @Expose
-    private Integer tutionIncome;
+    private double tutionIncome;
     @SerializedName("remittanceIncome")
     @Expose
-    private Integer remittanceIncome;
+    private double remittanceIncome;
     @SerializedName("interestIncomeOfFDR")
     @Expose
-    private Integer interestIncomeOfFDR;
+    private double interestIncomeOfFDR;
     @SerializedName("mobileNoId")
     @Expose
     private Integer mobileNoId;
@@ -355,7 +355,7 @@ public class CoApplicant {
         this.netSalaryType = netSalaryType;
     }
 
-    public Integer getNetSalary() {
+    public double getNetSalary() {
         return netSalary;
     }
 
@@ -363,39 +363,39 @@ public class CoApplicant {
         this.netSalary = netSalary;
     }
 
-    public Integer getBusinessIncome() {
+    public double getBusinessIncome() {
         return businessIncome;
     }
 
-    public void setBusinessIncome(Integer businessIncome) {
+    public void setBusinessIncome(double businessIncome) {
         this.businessIncome = businessIncome;
     }
 
-    public Integer getApartmentIncome() {
+    public double getApartmentIncome() {
         return apartmentIncome;
     }
 
-    public void setApartmentIncome(Integer apartmentIncome) {
+    public void setApartmentIncome(double apartmentIncome) {
         this.apartmentIncome = apartmentIncome;
     }
 
-    public Integer getSemipakaIncome() {
+    public double getSemipakaIncome() {
         return semipakaIncome;
     }
 
-    public void setSemipakaIncome(Integer semipakaIncome) {
+    public void setSemipakaIncome(double semipakaIncome) {
         this.semipakaIncome = semipakaIncome;
     }
 
-    public Integer getCommercialSpaceIncome() {
+    public double getCommercialSpaceIncome() {
         return commercialSpaceIncome;
     }
 
-    public void setCommercialSpaceIncome(Integer commercialSpaceIncome) {
+    public void setCommercialSpaceIncome(double commercialSpaceIncome) {
         this.commercialSpaceIncome = commercialSpaceIncome;
     }
 
-    public Integer getFactoryIncome() {
+    public double getFactoryIncome() {
         return factoryIncome;
     }
 
@@ -403,15 +403,15 @@ public class CoApplicant {
         this.factoryIncome = factoryIncome;
     }
 
-    public Integer getAgriculturalIncome() {
+    public double getAgriculturalIncome() {
         return agriculturalIncome;
     }
 
-    public void setAgriculturalIncome(Integer agriculturalIncome) {
+    public void setAgriculturalIncome(double agriculturalIncome) {
         this.agriculturalIncome = agriculturalIncome;
     }
 
-    public Integer getTutionIncome() {
+    public double getTutionIncome() {
         return tutionIncome;
     }
 
@@ -419,15 +419,15 @@ public class CoApplicant {
         this.tutionIncome = tutionIncome;
     }
 
-    public Integer getRemittanceIncome() {
+    public double getRemittanceIncome() {
         return remittanceIncome;
     }
 
-    public void setRemittanceIncome(Integer remittanceIncome) {
+    public void setRemittanceIncome(double remittanceIncome) {
         this.remittanceIncome = remittanceIncome;
     }
 
-    public Integer getInterestIncomeOfFDR() {
+    public double getInterestIncomeOfFDR() {
         return interestIncomeOfFDR;
     }
